LayoutTests/media/modern-media-controls/time-control/time-control-loading-indicator-layout-expected.txt

TimeControl activity indicator must occupy a flex slot when loading rather than overlapping the scrubber.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(timeControl.activityIndicator.element).position is "relative"
PASS timeControl.activityIndicator.element.getBoundingClientRect().right <= timeControl.scrubber.element.getBoundingClientRect().left is true
PASS successfullyParsed is true

TEST COMPLETE
